reorganizing Configuration folder
