+++ backend/src/Designer/Configuration/ServiceRepositorySettings.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using Altinn.Studio.Designer.Configuration.Marker;
 using Altinn.Studio.Designer.Helpers.Extensions;
 using Altinn.Studio.Designer.Models;
 using Newtonsoft.Json;
