+++ backend/src/Designer/Configuration/TestdataRepositorySettings.cs
@@ -1,4 +1,5 @@
 using System;
+using Altinn.Studio.Designer.Configuration.Marker;
 
 namespace Altinn.Studio.Designer.Configuration
 {
