+++ backend/src/Designer/Infrastructure/Models/PostgreSQLSettings.cs
@@ -1,3 +1,5 @@
+using Altinn.Studio.Designer.Configuration.Marker;
+
 namespace Altinn.Studio.Designer.Configuration
 {
     /// <summary>
